# bff/src/main/java/com/library/bff/dto/LoginRequest.java
package com.library.bff.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
	@Email @NotBlank String email,
	@NotBlank String password
) {
}