# import/openapi-swagger.json
{"openapi":"3.0.3","servers":[{"url":"http://localhost:5000","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/users/profile":{"put":{"tags":["user-controller"],"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserDto"}}}}}}},"/books/{id}":{"get":{"tags":["book-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBookResponse"}}}}}},"put":{"tags":["book-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBookResponse"}}}}}},"delete":{"tags":["book-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/books":{"get":{"tags":["book-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListBookResponse"}}}}}},"post":{"tags":["book-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBookResponse"}}}}}}},"/users":{"get":{"tags":["user-controller"],"operationId":"getAllProfiles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListUserDto"}}}}}}},"/books/search":{"get":{"tags":["book-controller"],"operationId":"search","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/BookSearchRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListBookResponse"}}}}}}},"/users/{id}":{"delete":{"tags":["user-controller"],"operationId":"deleteProfile","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}}},"components":{"schemas":{"ProfileUpdateRequest":{"type":"object","properties":{"fullName":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"}}},"ApiResponseUserDto":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/UserDto"},"message":{"type":"string"}}},"UserDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"authEmail":{"type":"string"},"fullName":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"}}},"BookRequest":{"required":["author","category","isbn","title"],"type":"object","properties":{"title":{"minLength":1,"type":"string"},"author":{"minLength":1,"type":"string"},"category":{"minLength":1,"type":"string"},"isbn":{"minLength":1,"type":"string"}}},"ApiResponseBookResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/BookResponse"},"message":{"type":"string"}}},"BookResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"author":{"type":"string"},"category":{"type":"string"},"isbn":{"type":"string"}}},"RegisterRequest":{"required":["email","password"],"type":"object","properties":{"email":{"minLength":1,"type":"string","format":"email"},"password":{"minLength":1,"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"minLength":1,"type":"string","format":"email"},"password":{"minLength":1,"type":"string"}}},"ApiResponseListUserDto":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}},"message":{"type":"string"}}},"ApiResponseListBookResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/BookResponse"}},"message":{"type":"string"}}},"BookSearchRequest":{"type":"object","properties":{"title":{"type":"string"},"author":{"type":"string"},"category":{"type":"string"},"isbn":{"type":"string"}}},"ApiResponseVoid":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object"},"message":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}