# ms-auth/src/main/java/com/library/auth/dto/LoginRequest.java
package com.library.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
	@Email @NotBlank String email,
	@NotBlank String password
) {
}