# ms-auth/src/main/java/com/library/auth/dto/RegisterRequest.java
package com.library.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record RegisterRequest(
	@Email @NotBlank String email,
	@NotBlank String password
) {
}